model: added 'Location', refactoring
Added the 'Location' class, like 'Profile'. As a specialty,
the Location is already prepared to hold geo-coordinates
from the mobile's GeoLocator services.

Did some refactoring to keep generic, common code in
GenericObservableList instead of each implementation.
This should later help when we actually start observing the lists and
persisting objects.